TUC-140 Refactored RK4 intake interval calculators
Now IntakeIntervalCalculator has two subclasses: one for analytical,
and one for RK4.
In IntakeIntervalCalculatorRK4, everything about Runge Kutta is implemented,
except the initial concentration values and the derivative.
These two functions have to be supplied by the subclasses.

The system uses the static inheritance pattern to allow a good optimization
of the RK4 process, with inline derivative functions.

	modified:   qtcreator/tucucore.pri
	modified:   ../src/tucucore/intakeintervalcalculator.cpp
	modified:   ../src/tucucore/intakeintervalcalculator.h
	new file:   ../src/tucucore/intakeintervalcalculatoranalytical.cpp
	new file:   ../src/tucucore/intakeintervalcalculatoranalytical.h
	new file:   ../src/tucucore/intakeintervalcalculatorrk4.cpp
	new file:   ../src/tucucore/intakeintervalcalculatorrk4.h
	modified:   ../src/tucucore/makefile
	modified:   ../src/tucucore/operation.cpp
	modified:   ../src/tucucore/pkmodels/onecompartmentbolus.h
	modified:   ../src/tucucore/pkmodels/onecompartmentextra.h
	modified:   ../src/tucucore/pkmodels/onecompartmentinfusion.h
	modified:   ../src/tucucore/pkmodels/rkonecompartmentextra.cpp
	modified:   ../src/tucucore/pkmodels/rkonecompartmentextra.h
	modified:   ../src/tucucore/pkmodels/rkonecompartmentgammaextra.cpp
	modified:   ../src/tucucore/pkmodels/rkonecompartmentgammaextra.h
	modified:   ../src/tucucore/pkmodels/threecompartmentbolus.h
	modified:   ../src/tucucore/pkmodels/threecompartmentextra.h
	modified:   ../src/tucucore/pkmodels/threecompartmentinfusion.h
	modified:   ../src/tucucore/pkmodels/twocompartmentbolus.h
	modified:   ../src/tucucore/pkmodels/twocompartmentextra.h
	modified:   ../src/tucucore/pkmodels/twocompartmentinfusion.h
	modified:   ../test/tucucore/pkmodels/constanteliminationbolus.h

+++ src/tucucore/operation.cpp
@@ -538,7 +538,7 @@ DynamicOperation::evaluate(const OperationInputList &_inputs, double &_result)
     int prefLevelBest = -1;
 
     // Find the 'best' operation to perform
-    for (int i = 0; i < static_cast<int>(m_operations.size()); ++i) {
+    for (size_t i = 0; i < m_operations.size(); ++i) {
         // Check that the given operation can run with the inputs passed as parameter
         if (m_operations.at(i).first->check(_inputs)) {
             const int nInputs = static_cast<int>(m_operations.at(i).first->getInputs().size());
@@ -547,7 +547,7 @@ DynamicOperation::evaluate(const OperationInputList &_inputs, double &_result)
             if (nInputs > nInputsBest || (nInputs == nInputsBest && prefLevel < prefLevelBest)) {
                 // We have found an operation with either more matching inputs or the same number but with higher
                 // preferability
-                idxBest = i;
+                idxBest = static_cast<int>(i);
                 nInputsBest = nInputs;
                 prefLevelBest = prefLevel;
             }
